faster copy times between the CPU and GPU Implementation in parahash.cu, based on parallelized_hash.cu.
Calling cudaMalloc and cudaMemcpy is extremely expensive, so we
can use structs with fixed size to avoid the double pointer indirection.
Furthermore, we get contiguous memory so it can be efficiently cached.

We can then use only 2 cudaMalloc and 2 cudaMemcpy.

This reduces the run time of the example from 90s to 0.3s.